fix not equal rule
